# Optimisation of Shannon Entropy function