--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyentrp"
-version = "0.8.2"
+version = "0.8.3"
 description = "A Python library for computing entropy measures for time series analysis."
 authors = ["Nikolay Donets <nd@donets.org>"]
 license = "apache-2.0"