--- a/setup.py
+++ b/setup.py
@@ -7,7 +7,7 @@ long_description = (this_directory / "README.md").read_text()
 
 setup(
     name="pyentrp",
-    version="0.8.2",
+    version="0.8.3",
     description="Functions on top of NumPy for computing different types of entropy",
     long_description=long_description,
     long_description_content_type="text/markdown",