- Version for `string` inputs uses `Counter` to improve performance
- Version for `non-string`
  - Uses `np.unique` with the `return_counts=True` parameter to directly obtain unique values and their corresponding counts from the time series array.
  - Utilizes NumPy's array operations to calculate frequencies and perform element-wise multiplication and summation without explicit loops.

Benchamrk:

Size | Original Time | Efficient Time (string) | Efficient Time (non-string) | Efficient (string) vs Original, % diff | Efficient (non-string) vs Original, % diff
-- | -- | -- | -- | -- | --
10 | 0.000839 | 0.000605 | 0.001009 | 27.89% | -20.26%
100 | 0.033200 | 0.004940 | 0.001059 | 85.12% | 96.81%
500 | 0.486557 | 0.022725 | 0.001692 | 95.33% | 99.65%
1000 | 1.882986 | 0.046219 | 0.002610 | 97.55% | 99.86%